RHINENG-18850: apply validateCanonicalFacts for all services
 - support canonical facts validation and error handlings for
   "compliance" and "malware-detection" services;
 - add metric counter "puptoo_messages_processed_failure_total"
